\name{lab5_shiny-package}
\alias{lab5_shiny-package}
\alias{lab5_shiny}
\docType{package}
\title{
\packageTitle{lab5_shiny}
}
\description{
\packageDescription{lab5_shiny}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lab5_shiny}
\packageIndices{lab5_shiny}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{lab5_shiny}

Maintainer: \packageMaintainer{lab5_shiny}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
